Crux.Interaction.ApplicationCommand.ChatInput (crux_interaction v0.1.0-dev)

Provides a DSL to define chat input application commands A.K.A. slash commands.

Restrictions

Names must consist only of lowercase characters (and -) and may be up to 32 characters long.
Descriptions may be up to 100 characters long without further restrictions.

These name and description restrictions also apply to options.
For choices the name and descriptions both may be up to 100 characters long without further restrictions.

A chat input application command may have up to

25 subcommand groups

25 subcommand

25 options

25 choices
per grouping / level.

Maximum of 4000 characters for combined name, description, and value properties for each command and its subcommands and groups.

It's recommend to checkout the Discord Developer Documentation on Chat Input Application Commands.

Examples

A basic ping command:

defmodule MyBot.Ping do
  use Crux.Interaction.ApplicationCommand.ChatInput

  @name "ping"
  @description "Pongs right back at you!"
end

A command using choices:

defmodule MyBot.Greet do
  use Crux.Interaction.ApplicationCommand.ChatInput

  @name "greet"
  @description "Greets a user"

  @required true
  string "type", "The type of the greeting" do
    choice "politely", "politely"
    choice "casually", "casually"
  end

  @required true
  user "user", "The user you'd like to greet"
end

A command using sub commands.
Sub command groups can be used to nest one level further:

defmodule MyBot.Info do
  use Crux.Interaction.ApplicationCommand.ChatInput
  
  @name "info"
  @description "Shows basic info about a user, channel, or role."
  # Only specific users / roles can use this command. (Needs to be separately specified per guild)
  @default_permission false

  subcommand "user", "Show basic info about a user" do
    @required true
    user "user", "The user you want to show info about"
  end

  subcommand "channel", "Show basic info about a channel" do
    @required true
    channel "channel", "The channel you want to show basic info about"
  end

  subcommand "role", "Show basic info about a role" do
    @required true
    role "role", "The role you want to show basic info about"
  end
end

A command using autocomplete to dynamically suggest the user options to chose from as they are typing:

defmodule MyBot.Tag do
  @moduledoc false
  use Crux.Interaction.ApplicationCommand.ChatInput

  @name "tag"
  @description "Shows a tag."

  @required true
  @autocomplete true
  string "name", "The name of the tag"
end

Functions

boolean(name, description)
(since 0.1.0) (macro)

Defines a boolean option, can optionally be @required true.

@required true
boolean "enabled", "Whether you want to enable or disable this."

channel(name, description)
(since 0.1.0) (macro)

Defines a channel option, can optionally be @required true.

@required true
channel "channel", "The channel you notifications to be sent to."

choice(name, value)
(since 0.1.0) (macro)

Defines choices for string/3, integer/3, and number/3 options.

Note that this is mutually exclusive with @autocomplete true.

string "foo", "foos" do
    choice "bar", "bar"
    choice "baz", "baz"
end

integer "foo", "foos" do
  choice "bar", 1
  choice "baz", 2
end

number "foo", "foos" do
  choice "bar", 1
  choice "baz", 1.5
end

integer(name, description, do_block \\ [do: nil])
(since 0.1.0) (macro)

Defines an integer option, can optionally be @required true and @autocomplete true.

@required true
integer "amount", "The amount of 🧀 you want to donate."

You can optionally define choices the user has to pick one from, see choice/2.

mentionable(name, description)
(since 0.1.0) (macro)

Defines a mentionable option, this is an union of user/2 and role/2, can optionally be @required true.

@required true
mentionable "target", "The user or role to block from this channel"

number(name, description, do_block \\ [do: nil])
(since 0.1.0) (macro)

Defines a number option, can optionally be @required true and @autocomplete true.

@required true
number "number", "Give me some number"

You can optionally define choices the user has to pick one from, see choice/2.

role(name, description)
(since 0.1.0) (macro)

Defines a role option, can optionally be @required true.

@required true
role "role", "The role you want to be mentioned alongside notifications."

string(name, description, do_block \\ [do: nil])
(since 0.1.0) (macro)

Defines a string option, can optionally be @required true and @autocomplete true.

@required true
string "name", "The name of the thing you want to create."

You can optionally define choices the user has to pick one from, see choice/2.

subcommand(name, description, do_block \\ [do: nil])
(since 0.1.0) (macro)

Defines a sub command option.

subcommand "create", "Creates... something!" do
  # Optionally further options
end

subcommand_group(name, description, do_block)
(since 0.1.0) (macro)

Defines a sub command group option.

subcommand_group "create", "Creates... something!" do
  # Required sub commands
end

user(name, description)
(since 0.1.0) (macro)

Defines a user option, can optionally be @required true.

@required true
user "user", "The user you want to interact with."