Crux.Interaction.ApplicationCommand.Exceptions.InvalidChoice exception (crux_interaction v0.1.0-dev)

Raised when an invalid application command chat input choice name or description was specified.
Also raised when too many or duplicated choices were specified.
See valid_name?/1 and valid_choice?/2

Functions

valid_value?(type, value)
(since 0.1.0)

Specs

valid_value?(type :: number(), value :: term()) :: boolean()

Whether the given value is valid for a chat input choice value.
Choice value validity depend on their enclosing option:

string/3 -> must be a [1,100] characters long string

integer/3 -> must be an integer in [-2^53,+2^53]

number/3 -> must be a double in [-2^53,+2^53]