Crux.Interaction.Component.SelectMenu (crux_interaction v0.1.0-dev)

Select menus are another interactive component that renders on messages.

For select menus the following limitations apply:

Select menus must be in action rows.

There may be one select menu in an action row.

You may also want to checkout the Discord Developer Documentation about select menus.

Types

t()

Specs

t() :: %Crux.Interaction.Component.SelectMenu{
  custom_id: String.t(),
  disabled: boolean() | nil,
  max_values: pos_integer() | nil,
  min_values: non_neg_integer() | nil,
  options: Crux.Interaction.Component.SelectMenu.Option.t(),
  placeholder: String.t() | nil,
  type: 3
}

Functions

add_options(select_menu, option)
(since 0.1.0)

Specs

add_options(t(), Crux.Interaction.Component.SelectMenu.Option.t()) :: t()

add_options(t(), [Crux.Interaction.Component.SelectMenu.Option.t()]) :: t()

Add one or multiple options to the select menu.

new(custom_id, option_s \\ [])
(since 0.1.0)

Specs

new(
  custom_id :: String.t(),
  option_s ::
    Crux.Interaction.Component.SelectMenu.Option.t()
    | [Crux.Interaction.Component.SelectMenu.Option.t()]
) :: t()

Create a new select menu with the given custom_id and option(s).

There is a maximum of 100 characters for the custom id and a limit of 25 options.

put_disabled(select_menu, disabled)
(since 0.1.0)

Specs

put_disabled(t(), boolean()) :: t()

Set whether the select menu is disabled.

put_max_values(select_menu, max_values)
(since 0.1.0)

Specs

put_max_values(t(), pos_integer()) :: t()

Set the maximum number of items that may be chosen.

put_min_values(select_menu, min_values)
(since 0.1.0)

Specs

put_min_values(t(), non_neg_integer()) :: t()

Set the minimum number of items that must be chosen.

put_placeholder(select_menu, placeholder)
(since 0.1.0)

Specs

put_placeholder(t(), String.t()) :: t()

Set a custom placeholder to be shown if nothing is selected.

There is a maximum of 100 characters allowed.