Crux.Interaction.Executor behaviour (crux_interaction v0.1.0-dev)

Behavior to provide commands and module to handle incoming interaction payloads to respond accordingly.

Usage

An example providing just registered commands:

defmodule MyApp.Executor do
  @behaviour Crux.Interaction.Executor

  @commands Crux.Interaction.Executor.map_commands([
    MyApp.Greet,
    MyApp.Info,
    MyApp.Ping,
    MyApp.Tag,
  ])

  @impl Crux.Interaction.Executor
  def get_commands() do
    @commands
  end

  # Optional callbacks here, as needed / wanted.
end

You could also not register any commands and just use the callbacks to handle interactions:

defmodule MyApp.Executor do
  @behaviour Crux.Interaction.Executor

  @impl Crux.Interaction.Executor
  def get_commands(), do: []

  @impl Crux.Interaction.Executor
  def handle_chat_input(interaction, context) do
    # Return some response
  end

  # Other optional callbacks here, as needed / wanted.
end

Or, of course, something in-between.

Plug Usage

You just need to plug your executor module into Crux.Interaction.Plug, see its moduledoc for a very small example. The plug will then take care of sending the response to Discord.

Gateway Usage

If you are using the gateway to receive interactions, you need to call execute/3 yourself and also send the response back yourself too.

That could look somewhat like this:

def handle_event(:INTERACTION_CREATE, interaction, _shard) do
  Crux.Interaction.Executor.execute(MyApp.Executor, interaction, %{some: :context}) # The context is optional and can be any term
  |> case do
    {:ok, response, _context} ->
      MyApp.Rest.create_interaction_response!(interaction.id, interaction.token, response)
    {:error, error_type, _context} ->
      # Do something with the error, see `t:error_type/0` for possible error types.
  end
end

Types

context()
(since 0.1.0)

Specs

context() :: term()

A context to pass to execute/3, which will be forwarded and maybe updated in handle_ callbacks.

error_type()
(since 0.1.0)

Specs

error_type() :: :unknown_type | :unknown_command | :unhandled_command

Available error types:

:unkown_type - The interaction type or sub-type (e.g. application commands) is unknown

:unhandled_type - The executor does not register an appropriate handler for this type of interaction (i.e. component / modal)

:unknown_command - The executor does not know about the given command

:unhandled_command - The executor does know about the given command, but the command doesn't register a handler

:no_return - A responsible handler was called, but it returned nil.

New types might be added in the future

handle_return()
(since 0.1.0)

Specs

handle_return() ::
  Crux.Rest.interaction_response()
  | {Crux.Rest.interaction_response(), new_context :: context()}
  | nil

Possible return values for the handle_ callbacks.

See also: reply/1

t()
(since 0.1.0)

Specs

t() :: module()

A module implementing the Crux.Interaction.Executor behaivor.

Callbacks

get_commands()
(since 0.1.0)

Specs

get_commands() :: %{
  required({name :: String.t(), type :: pos_integer()}) =>
    commands :: Crux.Interaction.ApplicationCommand.t()
}

Function returning all commands that the executor should work with.

See also map_commands/1 to map a list of Crux.Interaction.ApplicationCommand.t/0 to the expected return value.

handle_autocomplete(interaction, context)
(optional) (since 0.1.0)

Specs

handle_autocomplete(
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) :: handle_return()

Fallback function to handle incoming chat input application command interaction for commands that are not registered or that didn't register a Crux.Interaction.ApplicationCommand.handle_autocomplete/2 function.

handle_chat_input(interaction, context)
(optional) (since 0.1.0)

Specs

handle_chat_input(
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) :: handle_return()

Fallback function to handle incoming chat input application command interaction for commands that are not registered or that didn't register a Crux.Interaction.ApplicationCommand.handle/2 function.

handle_component(interaction, context)
(optional) (since 0.1.0)

Specs

handle_component(
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) :: handle_return()

Function to handle all incoming component interactions.

handle_interaction(interaction, context)
(optional) (since 0.1.0)

Specs

handle_interaction(
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) :: handle_return()

Fallback to handle incoming interactions of unknown (sub-)types.

handle_modal(interaction, context)
(optional) (since 0.1.0)

Specs

handle_modal(
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) :: handle_return()

Function to handle all incoming modal interactions.

Functions

execute(executor, interaction, context \\ %{})
(since 0.1.0)

Specs

execute(
  executor :: t(),
  interaction :: Crux.Interaction.interaction(),
  context :: context()
) ::
  {:ok, Crux.Rest.interaction_response(), new_context :: context()}
  | {:error, error_type(), new_context :: context()}

Handles an incoming interaction dispatching it to the relevant handle_ function either in this module or in one of the command modules.

If no suitable function could be determine an error response will be returned. For ping interactions this function will just return a pong response.

map_commands(commands)
(since 0.1.0)

Specs

map_commands([command :: Crux.Interaction.ApplicationCommand.t()]) :: %{
  required({name :: String.t(), type :: pos_integer()}) =>
    commands :: Crux.Interaction.ApplicationCommand.t()
}

Maps a list of commands to a map expected by get_commands/0.

You do not necessarily need to use this function, but it helps mapping them into the correct format.

reply(response)
(since 0.1.0)

Specs

reply(
  response ::
    Crux.Rest.interaction_response()
    | {Crux.Rest.interaction_response(), new_context :: context()}
) :: :ok | :error

Sends a response immediately. Useful / needed if the handler function needs more than 3 seconds to send a response or additional work after the initial reply is required.

Returns :error if a response was already sent, otherwise :ok.

Note that the handler function must return nil after using this method otherwise an error will be raised.