Crux.Interaction.MessageComponent.ActionRow (crux_interaction v0.1.0-dev)

An action row is a non-interactive container component for other types of components.

The allowed components within an action row depends on where its used.

For action rows the following limitations apply:

You can have at most 5 action rows within a message / interaction.

Action rows cannot contain action rows.

Action rows cannot mix components (as of yet).

You may also want to checkout the Discord Developer Documentation about action rows.

Types

t()
(since 0.1.0)

Specs

t() :: %Crux.Interaction.MessageComponent.ActionRow{
  components: [Crux.Interaction.MessageComponent.t()],
  type: 1
}

Functions

add_components(action_row, new_component)
(since 0.1.0)

Specs

add_components(t(), Crux.Interaction.MessageComponent.t()) :: t()

add_components(t(), [Crux.Interaction.MessageComponent.t()]) :: t()

Add one or multiple components to an action row.

new(component_s \\ [])
(since 0.1.0)

Specs

new(
  Crux.Interaction.MessageComponent.t()
  | [Crux.Interaction.MessageComponent.t()]
) :: t()

Create a new action row and optionally initialize it with the given component(s).