Crux.Interaction.MessageComponent.Button (crux_interaction v0.1.0-dev)

Buttons are interactive components.

For buttons the following limitations apply:

Buttons must be in action rows.

There may be up to 5 buttons per action row.

All buttons require at least one of label and emoji.

You may also want to checkout the Discord Developer Documentation about buttons.

Types

button_style()
(since 0.1.0)

Specs

button_style() :: 1..5

Available button styles:

primary - 1

secondary - 2

success - 3

danger - 4

link - 5

There is an image preview on the Discord Developer Documentation.

t()
(since 0.1.0)

Specs

t() :: %Crux.Interaction.MessageComponent.Button{
  custom_id: String.t() | nil,
  disabled: boolean() | nil,
  emoji:
    %{
      name: String.t() | nil,
      id: Crux.Structs.Snowflake.t() | nil,
      animated: boolean() | nil
    }
    | nil,
  label: String.t() | nil,
  style: button_style(),
  type: 2,
  url: String.t() | nil
}

Functions

button_style_danger()
(since 0.1.0)

Specs

button_style_danger() :: button_style()

The danger style, a red button.

button_style_link()
(since 0.1.0)

Specs

button_style_link() :: button_style()

The links style, a grey button with a "link" icon to signify its purpose.

button_style_primary()
(since 0.1.0)

Specs

button_style_primary() :: button_style()

The primary style, a blurple button.

button_style_secondary()
(since 0.1.0)

Specs

button_style_secondary() :: button_style()

The secondary style, a grey button.

button_style_success()
(since 0.1.0)

Specs

button_style_success() :: button_style()

The success style, a green button.

new(style, custom_id_or_url)
(since 0.1.0)

Specs

new(link_style :: button_style(), url :: String.t()) :: t()

new(non_link_style :: button_style(), custom_id :: String.t()) :: t()

Create a new button using the given style and custom_id / url.

See the button_style___ functions in this module.
There is a maximum of 100 characters for the custom_id (doesn't apply to the link style).

put_disabled(button, disabled)
(since 0.1.0)

Specs

put_disabled(t(), boolean()) :: t()

Set whether a button is disabled. (I.e. can't be clicked)

put_emoji(button, emoji)
(since 0.1.0)

Specs

put_emoji(
  t(),
  emoji :: %{
    name: String.t() | nil,
    id: Crux.Structs.Snowflake.t() | nil,
    animated: boolean() | nil
  }
) :: t()

Set the emoji for a button.

put_label(button, label)
(since 0.1.0)

Specs

put_label(t(), String.t()) :: t()

Set the label for a button.

There is a maximum of 80 characters.