Crux.Interaction.MessageComponent.InputText (crux_interaction v0.1.0-dev)

THIS IS NOT RELEASED / AVAILABLE YET.

EVERYTHING IS SUBJECT TO CHANGE.

For input text the following limitations apply:

Input text may only be used in an interaction response. (Not in regular messages)

Input text must be in action rows.

There may only be one input text per action row.

You may also want to checkout the Discord Developer Documentation about input text. (Does not exist yet)

Types

t()
(since 0.1.0)

Specs

t() :: %Crux.Interaction.MessageComponent.InputText{
  custom_id: String.t(),
  label: String.t(),
  max_length: pos_integer() | nil,
  min_length: non_neg_integer() | nil,
  placeholder: String.t() | nil,
  style: text_style(),
  type: 4
}

text_style()
(since 0.1.0)

Specs

text_style() :: 1..2

Availabel text styles:

short - 1

paragraph - 2

Functions

new(text_style, custom_id, label)
(since 0.1.0)

Specs

new(text_style(), custom_id :: String.t(), label :: String.t()) :: t()

Create an input text with the given custom_id and label.

put_max_length(input_text, max_length)
(since 0.1.0)

Specs

put_max_length(t(), pos_integer()) :: t()

Set the maximum length required for the text.

put_min_length(input_text, min_length)
(since 0.1.0)

Specs

put_min_length(t(), non_neg_integer()) :: t()

Set the minimum length required for the text.

put_placeholder(input_text, placeholder)
(since 0.1.0)

Specs

put_placeholder(t(), String.t()) :: t()

Set the placeholder text to show if no text was provided (yet).

text_style_paragraph()
(since 0.1.0)

The paragraph text syle, a multi line text box.

text_style_short()
(since 0.1.0)

The short text style, a single row for text.