Crux.Interaction.MessageComponent.SelectMenu.Option (crux_interaction v0.1.0-dev)

An option used in a select menu.

You may also want to checkout the Discord Developer Documentation about select menu options.

Types

t()
(since 0.1.0)

Specs

t() :: %Crux.Interaction.MessageComponent.SelectMenu.Option{
  default: boolean() | nil,
  description: String.t() | nil,
  emoji:
    %{
      name: String.t() | nil,
      id: Crux.Structs.Snowflake.t() | nil,
      animated: boolean() | nil
    }
    | nil,
  label: String.t(),
  value: String.t()
}

Functions

new(label, value)
(since 0.1.0)

Specs

new(label :: String.t(), value :: String.t()) :: t()

Create a new option using the given a label and value.

There is a 100 character limit for the label and a 100 character limit for the value.

put_default(option, default)
(since 0.1.0)

Specs

put_default(t(), boolean()) :: t()

Set whether this option will be selected by default.

put_description(option, description)
(since 0.1.0)

Specs

put_description(t(), String.t()) :: t()

Set the description shown below the option name.

put_emoji(option, emoji)
(since 0.1.0)

Specs

put_emoji(
  t(),
  emoji :: %{
    name: String.t() | nil,
    id: Crux.Structs.Snowflake.t() | nil,
    animated: boolean() | nil
  }
) :: t()

Set the emoji to display alongside the option name.