Crux.Interaction.Plug (crux_interaction v0.1.0-dev)

Wrapper plug around Crux.Interaction.Executor.

You need to supply an executor module as :executor to this plug.

  plug Crux.Interaction.Plug, executor: Your.Executor

Note that handler functions will receive a context with a conn key. If handler functions return a new context it must include that conn key with the (updated) conn struct.

Responds with 405 if the http verb is not POST.

In order to use this plug, you need to add plug to your dependencies. If you want to respond with files, you also need to add mimerl and hackney (for hackney_multipart) to your dependencies.

Functions

resolve_files(response)