Crux.Interaction.Plug.CacheBodyReader (crux_interaction v0.1.0-dev)

To be plugged into Plug.Parsers as a body_reader.

plug Plug.Parser,
  # ...
  body_reader: {Crux.Interaction.Plug.CacheBodyReader, :read_body, []},
  # ...

Caching the raw body prior to parsing it is necessary to be able to validate Discord's signature via Crux.Interaction.Plug.VerifyHeader.

The raw body will then be placed into conn.private[:raw_body].

Taken from https://hexdocs.pm/plug/Plug.Parsers.html#module-custom-body-reader and modified a bit

In order to use this module, you need to add plug to your dependencies.

Functions

read_cached_body(conn)
(since 0.1.0)

Specs

read_cached_body(Plug.Conn.t()) :: String.t() | nil

Get the cached body, returns nil if no body was read.