Crux.Interaction.Response (crux_interaction v0.1.0-dev)

Allows you to compose interaction responses.

First componse the response wanted through the with_ functions, then pass / pipe
the resulting data to the relevant function of the "Response Types" group.

Notes:

Multiple calls of the same with_ function will override the previous set value.

You can pass nil to one of the with_ functions to clear a previously set value.

Examples

# Send an ephemeral text response
with_content("Hello there")
|> with_flags(1 <<< 6)
|> channel_message()

# Update a message
with_content("Current state: #{state}")
|> update_message()

# Offer choices as response to an autocomplete interaction
with_choices([%{choice: "One", 1}, %{choice: "Two", 2}, %{choice: "Four", 4}])
|> autocomplete_result()

# Open a modal
# MODALS ARE NOT RELEASED / AVAILABLE YET.
with_custom_id("custom_id")
|> with_components(
  ActionRow.new(
    InputText.new(InputText.text_style_short(), "input_text_1", "What is your name?")
  )
)
|> modal()

Types

autocomplete_data()
(since 0.1.0)

Specs

autocomplete_data() :: %{choices: [choice()]}

Data used to respond to an autocomplete interaction.

Limits:

:choices - Max of 25 in length

For more information see the Discord Developer Documentation.

choice()
(since 0.1.0)

Specs

choice() :: %{name: String.t(), value: String.t() | integer() | float()}

A choice to be used in with_choices/1,2.

Limits:

:name - Length must be in [1,100].

:value - If string, length must be in [1,100], otherwise for integer and double value must be in [-2^53,+2^53].

For more information see the Discord Developer Documentation.

message_data()
(since 0.1.0)

Specs

message_data() :: %{
  tts: boolean(),
  content: String.t(),
  embeds: [Crux.Rest.embed_options()],
  allowed_mentions: %{
    parse: [roles_or_users_or_everyone :: String.t()],
    roles: [Crux.Structs.Snowflake.t()],
    users: [Crux.Structs.Snowflake.t()]
  },
  flags: non_neg_integer(),
  components: [Crux.Interaction.MessageComponent.t() | Crux.Rest.component()],
  attachments: [%{id: Crux.Structs.Snowflake.t()}]
}

Used when responding to an interaction with a message.

For more information see the Discord Developer Documentation.

modal_data()
(since 0.1.0)

Specs

modal_data() :: %{
  custom_id: String.t(),
  components: [Crux.Interaction.MessageComponent.t() | Crux.Rest.component()]
}

Used when responding to an interaction with a modal.

MODALS ARE NOT RELEASED / AVAILABLE YET.

For more information see the Discord Developer Documentation.

Response Types

autocomplete_result(data)
(since 0.1.0)

Specs

autocomplete_result(message_data() | modal_data()) ::
  Crux.Rest.interaction_response()

Generates a response that displays the user choices to chose from.

Only valid as a response to auto complete interactions.

channel_message(data)
(since 0.1.0)

Specs

channel_message(message_data() | modal_data()) ::
  Crux.Rest.interaction_response()

Generates a response that responds to an interaction with a message.

Not valid for autocomplete interactions.

deferred_channel_message()
(since 0.1.0)

Specs

deferred_channel_message() :: Crux.Rest.interaction_response()

Generates a response that acknowledges an interaction and displays a loading state to the user.
A response must be sent later manually.

Not valid for autocomplete interactions.

deferred_update_message()
(since 0.1.0)

Specs

deferred_update_message() :: Crux.Rest.interaction_response()

Generates a response that acknowledges a component interaction and displays no loading state to the user.
An update must be sent later manually.

Only valid as response to component interactions.

modal(data)
(since 0.1.0)

Specs

modal(modal_data()) :: Crux.Rest.interaction_response()

Generates a response that displays a modal to the user.

MODALS ARE NOT RELEASED / AVAILABLE YET.

Not valid as response to a modal interaction.

pong()
(since 0.1.0)

Specs

pong() :: Crux.Rest.interaction_response()

Generates a response that acknowledges a ping.

Only valid as response to Ping interactions.

update_message(data)
(since 0.1.0)

Specs

update_message(message_data()) :: Crux.Rest.interaction_response()

Generates a response that updates the message the component was attached to.

Only valid as response to component interactions.

Autocomplete

with_choices(data \\ %{}, choices)
(since 0.1.0)

Specs

with_choices(autocomplete_data(), [choice()]) :: autocomplete_data()

The choices to provide.

Modal

with_custom_id(data \\ %{}, custom_id)
(since 0.1.0)

Specs

with_custom_id(modal_data(), String.t()) :: modal_data()

The custom id to use.

MODALS ARE NOT RELEASED / AVAILABLE YET.

Message

with_allowed_mentions(data \\ %{}, allowed_mentions)
(since 0.1.0)

Specs

with_allowed_mentions(
  message_data(),
  allowed_mentions ::
    %{
      parse: [roles_or_users_or_everyone :: String.t()],
      roles: [Crux.Structs.Snowflake.t()],
      users: [Crux.Structs.Snowflake.t()]
    }
    | nil
) :: message_data()

Controls what (kind of) mentions to resolve from the content.

For more information see the Discord Developer Documentation.

with_attachments(data \\ %{}, attachments)
(since 0.1.0)

Specs

with_attachments(
  message_data(),
  attachments :: [%{id: Crux.Structs.Snowflake.t()}] | nil
) :: message_data()

The attachments to keep on the message.

Not setting this means that all attachments will be kept.
An empty list means that all attachments will be removed.

Only relevant when updating a message.

with_content(data \\ %{}, content)
(since 0.1.0)

Specs

with_content(message_data(), content :: String.t() | nil) :: message_data()

The message content to use.

with_embeds(data \\ %{}, embeds)
(since 0.1.0)

Specs

with_embeds(
  message_data(),
  embeds :: [embed :: Crux.Rest.embed_options()] | nil
) :: message_data()

The embeds to use, may be up to 10.

with_flags(data \\ %{}, flags)
(since 0.1.0)

Specs

with_flags(message_data(), flags :: non_neg_integer() | nil) :: message_data()

Message flags to use.

with_tts(data \\ %{}, tts)
(since 0.1.0)

Specs

with_tts(message_data(), tts :: boolean() | nil) :: message_data()

Whether the message content should use text-to-speech.

Multiple

with_components(data \\ %{}, components)
(since 0.1.0)

Specs

with_components(
  message_data(),
  Crux.Interaction.MessageComponent.t() | Crux.Rest.component()
) :: message_data()

with_components(
  modal_data(),
  Crux.Interaction.MessageComponent.t() | Crux.Rest.component()
) :: modal_data()

The components to use.

Available in modals and messages.